+++ src/transaction.ts
@@ -7,50 +7,24 @@ export let finalizationQueue: null | FinalizationQueue = null;
 /** Queue of effects to execute if the transaction succeeds. */
 export let effectQueue: EffectQueue = new Map();
 
-let runEffectsOnCommit = true;
-
 export const startTransaction = () => {
   finalizationQueue = new Map();
 };
 
-const collectEffects = () => {
-  const pendingEffects = effectQueue;
-  effectQueue = new Map();
-
-  return async () => {
-    await Promise.all([...pendingEffects.values()].map((effect) => effect()));
-  };
-};
-
-export const finishTransaction = (commit: boolean) => {
+export const finishTransaction = (
+  commit: boolean
+): (() => Promise<unknown>) => {
   finalizationQueue!.forEach((finalize) => finalize(commit));
   finalizationQueue = null;
 
-  if (runEffectsOnCommit) {
-    const runEffects = collectEffects();
+  const scheduledEffects = effectQueue;
+  effectQueue = new Map();
 
-    if (commit) {
-      // Errors should be handled as callbacks to the behavior. If the error
-      // bubbled this far, it's too late. Let it fail.
-      //
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      runEffects();
-    }
-  }
-};
-
-/**
- * Run an action, but suspend all effects until the returned callback is
- * invoked. Useful for testing.
- */
-export const suspendEffects = (callback: () => void): (() => Promise<void>) => {
-  try {
-    runEffectsOnCommit = false;
-    callback();
-    return collectEffects();
-  } finally {
-    runEffectsOnCommit = true;
-  }
+  return function runEffects() {
+    return Promise.all(
+      [...scheduledEffects.values()].map((effect) => effect())
+    );
+  };
 };
 
 type EffectQueue = Map<Behavior<unknown>, () => void | Promise<void>>;